string resources renamed
