+++ app/src/main/java/com/ably/tracking/demo/subscriber/ably/AssetTracker.kt
@@ -1,4 +1,4 @@
-package com.ably.tracking.demo.subscriber.domain
+package com.ably.tracking.demo.subscriber.ably
 
 import android.util.Log
 import com.ably.tracking.Accuracy
