+++ app/src/main/java/com/ably/tracking/demo/subscriber/ably/AssetTrackerAnimator.kt
@@ -1,4 +1,4 @@
-package com.ably.tracking.demo.subscriber.domain
+package com.ably.tracking.demo.subscriber.ably
 
 import com.ably.tracking.LocationUpdate
 import com.ably.tracking.ui.animation.LocationAnimator
